deletye main2.rb, move contents to main.rb to keep working oin main file from now on
